refactor: separated auctionFaqSection from FaqAndSpecialist section
