+++ src/app/Scenes/Artwork/Components/ArtworkExtraLinks/FaqAndSpecialistSection.tsx
@@ -1,35 +1,19 @@
 import { ArtworkExtraLinks_artwork } from "__generated__/ArtworkExtraLinks_artwork.graphql"
-import { AuctionTimerState } from "app/Components/Bidding/Components/Timer"
 import { navigate } from "app/navigation/navigate"
-import { partnerName } from "app/Scenes/Artwork/Components/ArtworkExtraLinks/partnerName"
 import { sendEmail } from "app/utils/sendEmail"
 import { Schema } from "app/utils/track"
-import { Sans, Spacer } from "palette"
+import { Sans } from "palette"
 import React from "react"
 import { Text } from "react-native"
 import { useTracking } from "react-tracking"
 
 interface FaqAndSpecialistSectionProps {
   artwork: ArtworkExtraLinks_artwork
-  auctionState: AuctionTimerState
 }
 
-export const FaqAndSpecialistSection: React.FC<FaqAndSpecialistSectionProps> = ({
-  artwork,
-  auctionState,
-}) => {
-  const { isAcquireable, isOfferable, isInAuction, sale, isForSale } = artwork
+export const FaqAndSpecialistSection: React.FC<FaqAndSpecialistSectionProps> = ({ artwork }) => {
+  const { isAcquireable, isOfferable } = artwork
   const { trackEvent } = useTracking()
-
-  const handleConditionsOfSaleTap = () => {
-    trackEvent(tracks.tappedConditionsOfSale())
-    navigate(`/conditions-of-sale`)
-  }
-
-  const handleReadOurAuctionFAQsTap = () => {
-    trackEvent(tracks.tappedAuctionFAQs())
-    navigate(`/auction-faq`)
-  }
 
   const handleAskASpecialistTap = (emailAddress: string) => {
     trackEvent(tracks.tappedAskASpecialist())
@@ -45,40 +29,7 @@ export const FaqAndSpecialistSection: React.FC<FaqAndSpecialistSectionProps> = (
     navigate(`/buy-now-feature-faq`)
   }
 
-  if (isInAuction && sale && isForSale && auctionState !== AuctionTimerState.CLOSED) {
-    return (
-      <>
-        <Sans size="2" color="black60">
-          By placing a bid you agree to {partnerName(sale)}{" "}
-          <Text
-            style={{ textDecorationLine: "underline" }}
-            onPress={() => handleConditionsOfSaleTap()}
-          >
-            Conditions of Sale
-          </Text>
-          .
-        </Sans>
-        <Spacer mb={1} />
-        <Sans size="2" color="black60">
-          Have a question?{" "}
-          <Text
-            style={{ textDecorationLine: "underline" }}
-            onPress={() => handleReadOurAuctionFAQsTap()}
-          >
-            Read our auction FAQs
-          </Text>{" "}
-          or{" "}
-          <Text
-            style={{ textDecorationLine: "underline" }}
-            onPress={() => handleAskASpecialistTap("specialist@artsy.net")}
-          >
-            ask a specialist
-          </Text>
-          .
-        </Sans>
-      </>
-    )
-  } else if (isAcquireable || isOfferable) {
+  if (isAcquireable || isOfferable) {
     return (
       <Sans size="2" color="black60">
         Have a question?{" "}
@@ -104,16 +55,6 @@ export const FaqAndSpecialistSection: React.FC<FaqAndSpecialistSectionProps> = (
 const tracks = {
   tappedAskASpecialist: () => ({
     action_name: Schema.ActionNames.AskASpecialist,
-    action_type: Schema.ActionTypes.Tap,
-    context_module: Schema.ContextModules.ArtworkExtraLinks,
-  }),
-  tappedAuctionFAQs: () => ({
-    action_name: Schema.ActionNames.AuctionsFAQ,
-    action_type: Schema.ActionTypes.Tap,
-    context_module: Schema.ContextModules.ArtworkExtraLinks,
-  }),
-  tappedConditionsOfSale: () => ({
-    action_name: Schema.ActionNames.ConditionsOfSale,
     action_type: Schema.ActionTypes.Tap,
     context_module: Schema.ContextModules.ArtworkExtraLinks,
   }),
